postgres/requirements_postgres.txt
flask==2.3.2
psycopg2-binary==2.9.6
gunicorn==20.1.0